services/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
requests
pandas
celery[redis]
redis
alembic
meilisearch

